make http-science a proxy
Before this change http-science was effectively a black hole--you sent
it requests, and it would always return 200 with an empty response. This
makes it somewhat difficult to deal with architecturally.

This change makes http-science proxy requests to CONTROL, i.e. it will
respond to requests with the response received from CONTROL. This makes
it easier to deal with architecturally, since it can be used as a
passthrough proxy.