Handle etag header
There are two cases to handle for etags:
1. The server can return etags for a request and this will be random by
server. This is simply solved by removing the etag before diffing the
responses.
2. If an etag was returned in a previous request the client can set the
Header 'If-None-Match' which tells the server to return a 304 if the
data hasn't changed. It isn't as obvious how to deal with this (see
comments in the code), but for now we've decided to ignore these
requests.