Update sfncli.mk
**Jira:**
https://clever.atlassian.net/browse/INFRANG-6510

**Overview:**
This is a combination of the go 1.21 upgrade, + sfncli and docker changes needed to avoid the glibc issue. This had to be rebased on master to avoid go.mod changes being overwritten and downgrading imports accidentally. 


**Testing:**
`make test` run locally and then deployed into private environment. After ensuring there were no logged errors and specifically looking for anything around glibc this is safe to merge. 


**Roll Out:**
Once this is merged and all other workers are merged I will be able to update sfncli master, and then go back and change the sfncli.mk code to grab "latest" again instead of the "pre-release".